requirements.txt
python-telegram-bot==20.1
openai>=0.27.0
python-dotenv==0.21.0
